requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.40.0
beautifulsoup4>=4.12.0
APScheduler>=3.10.0
pytest>=8.0.0
